# docs/swagger.yaml
basePath: /
definitions:
  model.CreatePost:
    properties:
      community_id:
        description: 社区ID*
        type: integer
      content:
        description: 内容
        type: string
      title:
        description: 标题*
        type: string
    required:
    - community_id
    - title
    type: object
  model.LoginGet:
    properties:
      password:
        description: 密码*
        type: string
      username:
        description: 用户名*
        type: string
    required:
    - password
    - username
    type: object
  model.ParamVoteDate:
    properties:
      direction:
        description: 投票 反对 取消 赞成
        example: "0"
        type: string
      post_id:
        description: 帖子id
        example: "0"
        type: string
    required:
    - post_id
    type: object
  model.RegisterForm:
    properties:
      confirm_password:
        description: re密码*
        type: string
      email:
        description: 邮箱*
        type: string
      password:
        description: 密码*
        type: string
      username:
        description: 用户名*
        type: string
      verification_code:
        description: 验证码*
        type: string
    required:
    - confirm_password
    - email
    - password
    - username
    - verification_code
    type: object
  model.ResponseStruct:
    properties:
      code:
        description: 业务状态码
        type: integer
      data:
        description: 数据
        type: object
      msg:
        description: 提示信息
        type: object
    type: object
host: Gin:8080
info:
  contact:
    email: 718428482@qq.com
    name: SongZhiBin
  description: 基于Grpc/Gin 扩展的一个社区系统
  termsOfService: Open source
  title: Happy
  version: "0.01"
paths:
  /CreatePost:
    post:
      consumes:
      - application/json
      description: 用于新建帖子接口 内部调用grpc接口
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 创建帖子参数
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/model.CreatePost'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ResponseStruct'
      security:
      - ApiKeyAuth: []
      summary: 新建帖子
      tags:
      - 帖子相关
  /GetPostDetail:
    get:
      consumes:
      - application/json
      description: 用于获取帖子详情接口 内部调用grpc接口
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 帖子ID*
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ResponseStruct'
      security:
      - ApiKeyAuth: []
      summary: 获取帖子详情
      tags:
      - 帖子相关
  /Login:
    post:
      consumes:
      - application/json
      description: 用于用户登录的接口 内部调用grpc接口
      parameters:
      - description: 登录参数
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/model.LoginGet'
      produces:
      - multipart/form-data
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ResponseStruct'
      security:
      - ApiKeyAuth: []
      summary: 登录
      tags:
      - 用户相关
  /PostList:
    get:
      consumes:
      - application/json
      description: 用于获取帖子列表接口 内部调用grpc接口
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 对应模式的ID(社区ID/用户ID)*
        in: query
        name: id
        required: true
        type: integer
      - description: 每页最大数量*
        in: query
        name: max
        type: integer
      - description: 查询模式*
        in: query
        name: mode
        type: integer
      - description: 对应页码*
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ResponseStruct'
      security:
      - ApiKeyAuth: []
      summary: 获取帖子列表
      tags:
      - 帖子相关
  /SignUp:
    post:
      consumes:
      - application/json
      description: 用于用户注册的接口 内部调用grpc接口
      parameters:
      - description: 注册参数
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/model.RegisterForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ResponseStruct'
      security:
      - ApiKeyAuth: []
      summary: 注册
      tags:
      - 用户相关
  /VerificationCode:
    post:
      consumes:
      - application/json
      description: 用于发送验证码的接口 内部调用grpc接口
      parameters:
      - description: 发送地址*
        in: query
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ResponseStruct'
      security:
      - ApiKeyAuth: []
      summary: 发送验证码
      tags:
      - 用户相关
  /Vote:
    post:
      consumes:
      - application/json
      description: 用于用户投票接口 内部调用grpc接口
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 获取帖子的模式
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/model.ParamVoteDate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ResponseStruct'
      security:
      - ApiKeyAuth: []
      summary: 用户投票接口
      tags:
      - 投票相关
  /communityDetail:
    get:
      consumes:
      - application/json
      description: 用于获取社区详情接口 内部调用grpc接口
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        type: string
      - description: 社区ID*
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ResponseStruct'
      security:
      - ApiKeyAuth: []
      summary: 获取社区详情
      tags:
      - 社区相关
  /communityList:
    get:
      consumes:
      - application/json
      description: 用于获取获取社区列表接口 内部调用grpc接口
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ResponseStruct'
      security:
      - ApiKeyAuth: []
      summary: 获取社区列表
      tags:
      - 社区相关
swagger: "2.0"